+++ docs/src/prototypes/chatPane/services/dataMock.ts
@@ -13,7 +13,6 @@ class ChatMock {
   private static readonly minUserCount = 2
   private static readonly defaultCount = 10
   private static readonly daysAgo = 40
-  private static readonly defaultChatTitle = 'Test Chat'
 
   private userIds: string[] = []
   private usersMap: Map<string, UserData> = new Map()
@@ -66,6 +65,7 @@ class ChatMock {
         timestamp: timestamp.short,
         timestampLong: timestamp.long,
         isImportant: random.boolean(),
+        withAttachment: random.boolean(),
       }
 
       return message
@@ -77,7 +77,7 @@ class ChatMock {
       isOneOnOne: this.usersMap.size === ChatMock.minUserCount,
       messages: this.chatMessages,
       members: this.usersMap,
-      title: ChatMock.defaultChatTitle,
+      title: `${currentUser.firstName} ${currentUser.lastName}`,
     }
   }