+++ docs/src/routes.tsx
@@ -6,12 +6,14 @@ import DocsLayout from './components/DocsLayout'
 import DocsRoot from './components/DocsRoot'
 
 import Accessibility from './views/Accessibility'
+import ColorPalette from './views/ColorPalette'
 import ShorthandProps from './views/ShorthandProps'
 import Introduction from './views/Introduction'
 import PageNotFound from './views/PageNotFound'
 import QuickStart from './views/QuickStart'
 import Theming from './views/Theming'
 import ThemingExamples from './views/ThemingExamples'
+import IntegrateCustomComponents from './views/IntegrateCustomComponents'
 
 const Router = () => (
   <BrowserRouter basename={__BASENAME__}>
@@ -52,11 +54,28 @@ const Router = () => (
             path="/prototype-search-page"
             component={require('./prototypes/SearchPage/index').default}
           />,
+          <DocsLayout
+            exact
+            path="/prototype-popups"
+            component={require('./prototypes/popups/index').default}
+          />,
+          <DocsLayout
+            exact
+            key="/icon-viewer"
+            path="/icon-viewer"
+            component={require('./prototypes/IconViewer/index').default}
+          />,
         ]}
         <DocsLayout exact path="/accessibility" component={Accessibility} />
         <DocsLayout exact path="/theming" component={Theming} />
         <DocsLayout exact path="/theming-examples" component={ThemingExamples} />
         <DocsLayout exact path="/shorthand-props" component={ShorthandProps} />
+        <DocsLayout
+          exact
+          path="/integrate-custom-components"
+          component={IntegrateCustomComponents}
+        />
+        <DocsLayout exact path="/color-palette" component={ColorPalette} />
         <DocsLayout exact path="/*" component={PageNotFound} />
       </Switch>
     </Switch>