+++ src/components/Chat/ChatItem.tsx
@@ -1,33 +1,22 @@
 import * as React from 'react'
-import * as PropTypes from 'prop-types'
 
-import { childrenExist, createShorthandFactory, RenderResultConfig, UIComponent } from '../../lib'
-import Slot from '../Slot/Slot'
-import { Extendable, ShorthandRenderFunction } from '../../../types/utils'
+import { Extendable, ShorthandValue } from '../../../types/utils'
 import {
+  childrenExist,
+  createShorthandFactory,
+  RenderResultConfig,
+  UIComponent,
   UIComponentProps,
   ChildrenComponentProps,
   ContentComponentProps,
-} from '../../lib/commonPropInterfaces'
-import {
-  commonUIComponentPropTypes,
-  childrenComponentPropTypes,
-  contentComponentPropsTypes,
-} from '../../lib/commonPropTypes'
+  commonPropTypes,
+} from '../../lib'
+import Slot from '../Slot/Slot'
 
 export interface ChatItemProps
-  extends UIComponentProps<any, any>,
+  extends UIComponentProps,
     ChildrenComponentProps,
-    ContentComponentProps {
-  /**
-   * A custom render function the content slot.
-   *
-   * @param {React.ReactType} Component - The computed component for this slot.
-   * @param {object} props - The computed props for this slot.
-   * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
-   */
-  renderContent?: ShorthandRenderFunction
-}
+    ContentComponentProps<ShorthandValue> {}
 
 /**
  * A chat item represents a single event in a chat.
@@ -40,10 +29,9 @@ class ChatItem extends UIComponent<Extendable<ChatItemProps>, any> {
   static displayName = 'ChatItem'
 
   static propTypes = {
-    ...commonUIComponentPropTypes,
-    ...childrenComponentPropTypes,
-    ...contentComponentPropsTypes,
-    renderContent: PropTypes.func,
+    ...commonPropTypes.createCommon({
+      content: 'shorthand',
+    }),
   }
 
   static defaultProps = {
@@ -57,7 +45,7 @@ class ChatItem extends UIComponent<Extendable<ChatItemProps>, any> {
     variables,
     rest,
   }: RenderResultConfig<ChatItemProps>) {
-    const { children, content, renderContent } = this.props
+    const { children, content } = this.props
 
     return (
       <ElementType {...rest} className={classes.root}>
@@ -66,7 +54,6 @@ class ChatItem extends UIComponent<Extendable<ChatItemProps>, any> {
           : Slot.create(content, {
               styles: styles.content,
               variables: variables.content,
-              render: renderContent,
             })}
       </ElementType>
     )