+++ src/components/RadioGroup/RadioGroup.tsx
@@ -4,22 +4,20 @@ import * as _ from 'lodash'
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
-import { AutoControlledComponent, childrenExist, customPropTypes } from '../../lib'
+import {
+  AutoControlledComponent,
+  childrenExist,
+  customPropTypes,
+  UIComponentProps,
+  ChildrenComponentProps,
+  commonPropTypes,
+} from '../../lib'
 import RadioGroupItem, { RadioGroupItemProps } from './RadioGroupItem'
 import { radioGroupBehavior } from '../../lib/accessibility'
 import { Accessibility, AccessibilityActionHandlers } from '../../lib/accessibility/types'
+import { Extendable, ShorthandValue, ComponentEventHandler } from '../../../types/utils'
 
-import {
-  Extendable,
-  ShorthandValue,
-  ShorthandRenderFunction,
-  ComponentEventHandler,
-} from '../../../types/utils'
-import { UIComponentProps, ChildrenComponentProps } from '../../lib/commonPropInterfaces'
-import { commonUIComponentPropTypes, childrenComponentPropTypes } from '../../lib/commonPropTypes'
-
-// TODO check if the checkedValueChanged args are correct
-export interface RadioGroupProps extends UIComponentProps<any, any>, ChildrenComponentProps {
+export interface RadioGroupProps extends UIComponentProps, ChildrenComponentProps {
   /**
    * Accessibility behavior if overridden by the user.
    * @default radioGroupBehavior
@@ -42,16 +40,6 @@ export interface RadioGroupProps extends UIComponentProps<any, any>, ChildrenCom
   /** Shorthand array of props for RadioGroup. */
   items?: ShorthandValue[]
 
-  /**
-   * A custom render iterator for rendering each of the RadioGroup items.
-   * The default component, props, and children are available for each item.
-   *
-   * @param {React.ReactType} Component - The computed component for this slot.
-   * @param {object} props - The computed props for this slot.
-   * @param {ReactNode|ReactNodeArray} children - The computed children for this slot.
-   */
-  renderItem?: ShorthandRenderFunction
-
   /** A vertical radio group displays elements vertically. */
   vertical?: boolean
 }
@@ -67,14 +55,14 @@ class RadioGroup extends AutoControlledComponent<Extendable<RadioGroupProps>, an
   static create: Function
 
   static propTypes = {
-    ...commonUIComponentPropTypes,
-    ...childrenComponentPropTypes,
+    ...commonPropTypes.createCommon({
+      content: false,
+    }),
     accessibility: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
     checkedValue: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     defaultCheckedValue: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     items: customPropTypes.collectionShorthand,
     checkedValueChanged: PropTypes.func,
-    renderItem: PropTypes.func,
     vertical: PropTypes.bool,
   }
 
@@ -172,13 +160,12 @@ class RadioGroup extends AutoControlledComponent<Extendable<RadioGroupProps>, an
   })
 
   private renderItems = (vertical: boolean) => {
-    const { items, renderItem } = this.props
+    const { items } = this.props
 
     return _.map(items, item =>
       RadioGroupItem.create(item, {
         defaultProps: { vertical },
         overrideProps: this.handleItemOverrides,
-        render: renderItem,
       }),
     )
   }