+++ src/components/Slot/Slot.tsx
@@ -1,20 +1,17 @@
 import * as React from 'react'
-import { childrenExist, createShorthandFactory } from '../../lib'
 import {
+  childrenExist,
+  createShorthandFactory,
   UIComponentProps,
   ContentComponentProps,
   ChildrenComponentProps,
-} from '../../lib/commonPropInterfaces'
-import {
-  commonUIComponentPropTypes,
-  contentComponentPropsTypes,
-  childrenComponentPropTypes,
-} from '../../lib/commonPropTypes'
+  commonPropTypes,
+} from '../../lib'
 import createComponent, { CreateComponentReturnType } from '../../lib/createComponent'
 import * as PropTypes from 'prop-types'
 
 export interface SlotProps
-  extends UIComponentProps<SlotProps, any>,
+  extends UIComponentProps<SlotProps>,
     ContentComponentProps,
     ChildrenComponentProps {
   /**
@@ -36,9 +33,7 @@ const Slot: CreateComponentReturnType<SlotProps> & {
   className: 'ui-slot',
 
   propTypes: {
-    ...commonUIComponentPropTypes,
-    ...contentComponentPropsTypes,
-    ...childrenComponentPropTypes,
+    ...commonPropTypes.createCommon(),
     innerRef: PropTypes.func,
   },