+++ src/index.ts
@@ -108,7 +108,9 @@ export { default as Status, StatusPropsWithDefaults, StatusProps } from './compo
 
 export { default as Text, TextProps } from './components/Text/Text'
 
-export { default as Transition, TransitionProps } from './components/Transition/Transition'
+export { default as Animation, AnimationProps } from './components/Animation/Animation'
+
+export { default as Tree } from './components/Tree'
 
 //
 // Accessibility
@@ -140,3 +142,15 @@ export { default as dialogBehavior } from './lib/accessibility/Behaviors/Dialog/
 // Utilities
 //
 export { default as mergeThemes } from './lib/mergeThemes'
+export { createComponent } from './lib'
+export {
+  RenderStardustResultConfig,
+  CreateStardustComponentConfig,
+} from './lib/createStardustComponent'
+export {
+  AnimatedComponentProps,
+  StyledComponentProps,
+  UIComponentProps,
+  ChildrenComponentProps,
+  ContentComponentProps,
+} from './lib'