+++ src/themes/teams/componentStyles.ts
@@ -55,4 +55,8 @@ export { default as Status } from './components/Status/statusStyles'
 
 export { default as Text } from './components/Text/textStyles'
 
-export { default as Transition } from './components/Transition/transitionStyles'
+export { default as Tree } from './components/Tree/treeStyles'
+export { default as TreeItem } from './components/Tree/treeItemStyles'
+export { default as TreeTitle } from './components/Tree/treeTitleStyles'
+
+export { default as Animation } from './components/Animation/animationStyles'