+++ src/themes/teams/componentVariables.ts
@@ -49,4 +49,6 @@ export { default as Status } from './components/Status/statusVariables'
 
 export { default as Text } from './components/Text/textVariables'
 
-export { default as Transition } from './components/Transition/transitionVariables'
+export { default as TreeTitle } from './components/Tree/treeTitleVariables'
+
+export { default as Animation } from './components/Animation/animationVariables'