+++ src/themes/teams/components/Divider/dividerStyles.ts
@@ -1,3 +1,5 @@
+import * as _ from 'lodash'
+
 import { childrenExist, pxToRem } from '../../../../lib'
 import { ComponentSlotStylesInput, ICSSInJSStyle, ICSSPseudoElementStyle } from '../../../types'
 import { DividerPropsWithDefaults } from '../../../../components/Divider/Divider'
@@ -7,10 +9,13 @@ const dividerBorderStyle = (size, color): ICSSInJSStyle => ({
   background: color,
 })
 
-const beforeAndAfter = (size, type, variables): ICSSPseudoElementStyle => ({
+const beforeAndAfter = (color, size, type, variables): ICSSPseudoElementStyle => ({
   content: '""',
   flex: 1,
   ...dividerBorderStyle(size, variables.dividerColor),
+  ...(color && {
+    ...dividerBorderStyle(size, _.get(variables.colors, [color, 500])),
+  }),
   ...(type === 'primary' && {
     ...dividerBorderStyle(size, variables.primaryColor),
   }),
@@ -18,7 +23,7 @@ const beforeAndAfter = (size, type, variables): ICSSPseudoElementStyle => ({
 
 const dividerStyles: ComponentSlotStylesInput<DividerPropsWithDefaults, any> = {
   root: ({ props, variables }): ICSSInJSStyle => {
-    const { children, fitted, size, type, important, content } = props
+    const { children, color, fitted, size, type, important, content } = props
     return {
       color: variables.textColor,
       display: 'flex',
@@ -26,6 +31,9 @@ const dividerStyles: ComponentSlotStylesInput<DividerPropsWithDefaults, any> = {
       ...(!fitted && {
         paddingTop: variables.dividerPadding,
         paddingBottom: variables.dividerPadding,
+      }),
+      ...(color && {
+        color: _.get(variables.colors, [color, 500]),
       }),
       ...(type === 'primary' && {
         color: variables.primaryColor,
@@ -39,17 +47,17 @@ const dividerStyles: ComponentSlotStylesInput<DividerPropsWithDefaults, any> = {
             fontSize: pxToRem(12 + size),
             lineHeight: variables.textLineHeight,
             '::before': {
-              ...beforeAndAfter(size, type, variables),
+              ...beforeAndAfter(color, size, type, variables),
               marginRight: pxToRem(20),
             },
             '::after': {
-              ...beforeAndAfter(size, type, variables),
+              ...beforeAndAfter(color, size, type, variables),
               marginLeft: pxToRem(20),
             },
           }
         : {
             '::before': {
-              ...beforeAndAfter(size, type, variables),
+              ...beforeAndAfter(color, size, type, variables),
             },
           }),
     }