+++ src/themes/types.ts
@@ -11,6 +11,82 @@ import { Extendable, ObjectOf, ObjectOrFunc } from '../../types/utils'
 // We use these terms in typings to indicate which phase the typings apply to.
 
 // ========================================================
+// Colors
+// ========================================================
+
+/**
+ * A type for a palette for a single color.
+ */
+export type ColorVariants = {
+  50: string
+  100: string
+  200: string
+  300: string
+  400: string
+  500: string
+  600: string
+  700: string
+  800: string
+  900: string
+}
+
+/**
+ * A type for a predefined natural colors.
+ */
+type NaturalColorsStrict = Partial<{
+  blue: ColorVariants
+  green: ColorVariants
+  grey: ColorVariants
+  orange: ColorVariants
+  pink: ColorVariants
+  purple: ColorVariants
+  teal: ColorVariants
+  red: ColorVariants
+  yellow: ColorVariants
+}>
+
+export type NaturalColors = Extendable<NaturalColorsStrict, ColorVariants>
+
+/**
+ * A type for a predefined state colors.
+ */
+export type ContextualColorsStrict = Partial<{
+  text: ColorVariants
+
+  danger: ColorVariants
+  info: ColorVariants
+  success: ColorVariants
+  warning: ColorVariants
+}>
+
+export type ContextualColors = Extendable<ContextualColorsStrict, ColorVariants>
+
+/**
+ * A type for a predefined emphasis colors.
+ */
+type EmphasisColorsStrict = Partial<{
+  primary: ColorVariants
+  secondary: ColorVariants
+}>
+
+export type EmphasisColors = Extendable<EmphasisColorsStrict, ColorVariants>
+
+/**
+ * A type for a base colors.
+ */
+export type PrimitiveColors = Partial<{
+  black: string
+  white: string
+}>
+
+type ExtendablePalette<T> = T &
+  { [K in keyof T]?: K extends keyof PrimitiveColors ? string : ColorVariants }
+
+export type ColorPalette = ExtendablePalette<
+  EmphasisColorsStrict & ContextualColorsStrict & NaturalColorsStrict & PrimitiveColors
+>
+
+// ========================================================
 // Props
 // ========================================================
 
@@ -30,11 +106,19 @@ export type State = ObjectOf<any>
 // ========================================================
 
 export interface SiteVariablesInput extends ObjectOf<any> {
+  colors?: ColorPalette
+  contextualColors?: ContextualColors
+  emphasisColors?: EmphasisColors
+  naturalColors?: NaturalColorsStrict
   brand?: string
   htmlFontSize?: string
 }
 
 export interface SiteVariablesPrepared extends ObjectOf<any> {
+  colors?: ColorPalette
+  contextualColors?: ContextualColors
+  emphasisColors?: EmphasisColors
+  naturalColors?: NaturalColorsStrict
   brand?: string
   htmlFontSize?: string
   fontSizes: ObjectOf<string>
@@ -115,7 +199,7 @@ export interface ComponentSlotStylesPrepared<TProps = {}, TVars = {}>
 export interface ComponentSlotClasses extends ObjectOf<string> {}
 export interface ComponentSlotClasses extends ObjectOf<string> {}
 
-export type Animation =
+export type AnimationProp =
   | {
       name: string
       delay?: string
@@ -191,6 +275,7 @@ export interface ThemeComponentStylesInput {
   [key: string]: ComponentSlotStylesInput | undefined
 
   Accordion?: ComponentSlotStylesInput
+  Animation?: ComponentSlotStylesInput
   Attachment?: ComponentSlotStylesInput
   Avatar?: ComponentSlotStylesInput
   Button?: ComponentSlotStylesInput
@@ -230,6 +315,7 @@ export interface ThemeComponentStylesPrepared {
   [key: string]: ComponentSlotStylesPrepared | undefined
 
   Accordion?: ComponentSlotStylesPrepared
+  Animation?: ComponentSlotStylesPrepared
   Attachment?: ComponentSlotStylesPrepared
   Avatar?: ComponentSlotStylesPrepared
   Button?: ComponentSlotStylesPrepared
@@ -269,6 +355,7 @@ export interface ThemeComponentVariablesInput {
   [key: string]: any
 
   Accordion?: ComponentVariablesInput
+  Animation?: ComponentVariablesInput
   Attachment?: ComponentVariablesInput
   Avatar?: ComponentVariablesInput
   Button?: ComponentVariablesInput
@@ -307,6 +394,7 @@ export interface ThemeComponentVariablesPrepared {
   [key: string]: any
 
   Accordion?: ComponentVariablesPrepared
+  Animation?: ComponentVariablesPrepared
   Attachment?: ComponentVariablesPrepared
   Avatar?: ComponentVariablesPrepared
   Button?: ComponentVariablesPrepared