+++ test/jest.config.strict.js
@@ -1,0 +1,6 @@
+const commonConfig = require('./jest.config.common')
+
+module.exports = Object.assign({}, 
+    commonConfig, 
+    { setupTestFrameworkScriptFile: `${__dirname}/setup.strict.ts` }
+)