+++ test/specs/behaviors/testHelper.tsx
@@ -1,7 +1,7 @@
 import * as _ from 'lodash'
 import { Accessibility } from '../../../src/lib/accessibility/types'
 
-interface FilteredDescription {
+interface FilteredSpecification {
   behaviorName: string
   testMethod: (arg: TestMethod) => void
   params: RegExpExecArray
@@ -17,11 +17,18 @@ export interface TestDefinition {
   testMethod: (arg: TestMethod) => void
 }
 
+const skipSpecChecksForFiles = [
+  'chatBehavior.ts', // issue https://github.com/stardust-ui/react/issues/476
+  'chatMessageBehavior.ts', // issue https://github.com/stardust-ui/react/issues/476
+  'listBehavior.ts', // tests are written in listBehavior-test.tsx
+  'listItemBehavior.ts', // tests are written in listItemBehavior-test.tsx
+]
+
 export class TestHelper {
   private behaviors: Map<string, Accessibility> = new Map<string, Accessibility>()
   private testDefinitions: TestDefinition[] = []
 
-  private filteredDescriptionWithAssignedTestMethod: FilteredDescription[] = []
+  private filteredSpecificationWithAssignedTestMethod: FilteredSpecification[] = []
 
   public addBehavior(name: string, behavior: Accessibility) {
     this.behaviors.set(name, behavior)
@@ -40,7 +47,7 @@ export class TestHelper {
   public run(behaviorMenuItems: any) {
     this.findRegexAndAssingCorrespondingInfoToArray(behaviorMenuItems)
 
-    const groupedByBehavior = _(this.filteredDescriptionWithAssignedTestMethod)
+    const groupedByBehavior = _(this.filteredSpecificationWithAssignedTestMethod)
       .groupBy('behaviorName')
       .value()
     _.each(groupedByBehavior, (value, key) => {
@@ -60,40 +67,75 @@ export class TestHelper {
   public findRegexAndAssingCorrespondingInfoToArray(behaviorMenuItems: any) {
     behaviorMenuItems.forEach(behavior => {
       behavior.variations.forEach(variant => {
-        variant.text.split('\n').forEach(singleLineText => {
-          this.iterateRegexDefinitions(singleLineText, variant.name)
-        })
+        if (!variant.specification && !variant.description) {
+          this.failDescriptionPresenceTest(variant.name)
+        }
+        if (!variant.specification && !skipSpecChecksForFiles.find(item => item === variant.name)) {
+          this.failSpecificationPresenceTest(variant.name)
+        } else {
+          variant.specification.split('\n').forEach(specLine => {
+            if (specLine) {
+              this.iterateRegexDefinitions(specLine, variant.name)
+            }
+          })
+        }
       })
     })
   }
 
-  public iterateRegexDefinitions(singleLineText: string, behaviorName: string) {
+  public iterateRegexDefinitions(specLine: string, behaviorName: string) {
+    let regexMatched = false
     this.testDefinitions.forEach(testDefinition => {
       const regex = new RegExp(testDefinition.regexp)
-      const result = regex.exec(singleLineText)
+      const result = regex.exec(specLine)
       if (result) {
-        this.filteredDescriptionWithAssignedTestMethod.push({
+        regexMatched = true
+        this.filteredSpecificationWithAssignedTestMethod.push({
           behaviorName,
           testMethod: testDefinition.testMethod,
           params: result,
         })
       }
     })
+    if (!regexMatched) {
+      test(`${behaviorName} \n LINE: ${specLine} `, () => {
+        fail(
+          `Line mentioned in **behavior specification** doesn't match any regex expression of validation tests.`,
+        )
+      })
+    }
   }
 
   public getBehavior(behaviorName: string): Accessibility {
     const baseBehaviorName = behaviorName.replace('.ts', '')
     const importedBehavior = this.behaviors.get(baseBehaviorName)
     if (!importedBehavior) {
-      throw 'Behavior file was not found, probably was not imported. Import file and add behavior.'
+      throw 'Accessibility behavior file was not found, probably was not imported. Import file and add behavior.'
     }
     return importedBehavior
   }
 
-  public convertToBooleanIfApplicable(stringToConvert: string) {
-    if (stringToConvert === 'true' || stringToConvert === 'false') {
-      return Boolean(stringToConvert)
+  public convertToBooleanIfApplicable(stringToConvert: any) {
+    if (stringToConvert === 'true') {
+      return true
+    }
+    if (stringToConvert === 'false') {
+      return false
     }
     return stringToConvert
+  }
+
+  private failSpecificationPresenceTest(behaviorFileName: string) {
+    test(`${behaviorFileName} : Accessibility behavior is missing specification tag.`, () => {
+      fail(
+        `Accessibility behavior should have specification tag. If tests are written in separate file then add behavior file name into 'skipSpecChecksForFiles'.`,
+      )
+    })
+  }
+
+  private failDescriptionPresenceTest(behaviorFileName: string) {
+    test(`${behaviorFileName} : Accessibility behavior is missing description.`, () => {
+      fail('Accessibility behavior should have description.')
+    })
   }
 }